requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.1